+++ src/views/Workload/CreateView.vue
@@ -38,7 +38,7 @@
   import { useRoute } from 'vue-router'
   import { handleTrackerError } from '@/utils/errorHandlingTracker'
   import { workloadService } from '@/services/v2'
-  import { validationSchema } from './composables/validation'
+  import { validationSchema } from './configForm/validation'
 
   /**@type {import('@/plugins/analytics/AnalyticsTrackerAdapter').AnalyticsTrackerAdapter} */
   const tracker = inject('tracker')
